feat: mock Fiber for cli and rspec
